Moved canBeAttachedToCase to CallbackValidator
